project/requirements.txt
flask
sys
flask_restful
fileinput
mysql.connector
datetime
json